# Improve performance of fuzzy search when having a huge search index