@Toflar 2020-07-06
True, we'd need to store the regular words in `tl_search_terms` and the fuzzy variants in `tl_search_terms_fuzzy` or whatever. Wdyt, @ausi?