@ausi 2020-07-06
I think the problem is not that searching for `%e%` is slow. This takes about 7ms with my test data while `e%` takes about 4ms.

The problem is that `%e%` returns a very large set of rows that gets then merged with the very, very large `tl_search_index` table.

To improve fuzzy search we should create an optimized index for one- two- and three-letter wildcards that only correspond to a single `tl_search_index` row. This would result in about 10,000 rows after the JOIN for `%e%` instead of 1,662,000 rows. This would make searching for `%e%` basically as fast as searching for `%e%`.

But I’m not sure how far we want to go with the “old” search implementation.