@Toflar 2020-07-07
After further discussions with @ausi I'm going to close this because it's not the right approach. The problem isn't really searching the terms, it's the fact that there's too many joins to make after that. This can only be fixed by adding stemming and decomposition to the search indexing process so we don't even need wildcard search anymore.
But this will be for another day.